pyproject.toml
[project]
name = "portfolio-builder"
version = "0.1.0"
description = "An AI-powered portfolio builder using OpenAI Agent SDK with handoff architecture"
readme = "README.md"
authors = [
    { name = "waleed260", email = "vkdeku20@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "rich>=14.3.3",
]

[project.scripts]
portfolio-builder = "portfolio_builder:main"

[build-system]
requires = ["uv_build>=0.9.25,<0.10.0"]
build-backend = "uv_build"
